refactor: add effect registry and modular handlers
